plots indexed solution_mat
